@@ -10,63 +13,45 @@ dependencies {
     implementation project(path: ':afero-sdk-core')
 
     // https://github.com/square/retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.3.0'
-    implementation 'com.squareup.retrofit2:converter-jackson:2.3.0'
-    implementation 'com.squareup.retrofit2:adapter-rxjava:2.3.0'
-    implementation 'com.squareup.retrofit2:retrofit-mock:2.3.0'
+    implementation 'com.squareup.retrofit2:retrofit:3.0.0'
+    implementation 'com.squareup.retrofit2:converter-jackson:3.0.0'
+    implementation 'com.squareup.retrofit2:adapter-rxjava:3.0.0'
+    implementation 'com.squareup.retrofit2:retrofit-mock:3.0.0'
 
     // https://github.com/square/okhttp
-    implementation 'com.squareup.okhttp3:okhttp:3.8.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:3.8.0'
+    implementation 'com.squareup.okhttp3:okhttp:5.3.0'
+    implementation 'com.squareup.okhttp3:logging-interceptor:5.3.0'
 
     // https://github.com/FasterXML/jackson
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.8.5'
-    implementation 'com.fasterxml.jackson.core:jackson-core:2.8.5'
-    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.8.5'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.20.1'
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.20.1'
+    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.20'
 
     // https://github.com/ReactiveX/RxJava
-    implementation 'io.reactivex:rxjava:1.3.0'
+    implementation 'io.reactivex:rxjava:1.3.8'
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_7
-targetCompatibility = JavaVersion.VERSION_1_7
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
 
-publishing {
-    publications {
-        jar(MavenPublication) {
-            version = rootProject.version
-            artifactId project.getName()
-            groupId rootProject.group
-
-            artifact("$buildDir/libs/${project.getName()}.jar")
-//            artifact packageJavadoc
-
-            pom.withXml {
-                def dependencies = asNode().appendNode('dependencies')
-                configurations.getByName("compile").getResolvedConfiguration().getFirstLevelModuleDependencies().each {
-                    // don't include local dependencies
-                    if (it.moduleVersion != 'unspecified') {
-                        def dependency = dependencies.appendNode('dependency')
-                        dependency.appendNode('groupId', it.moduleGroup)
-                        dependency.appendNode('artifactId', it.moduleName)
-                        dependency.appendNode('version', it.moduleVersion)
-                    }
-                }
+afterEvaluate {
+    publishing {
+        publications {
+	    release(MavenPublication) { 
+	        from components.java
+		groupId rootProject.group
+		version = rootProject.version
+		artifactId project.name
             }
-        }
+	}
+	repositories {
+	    maven {
+	        url = 'https://afero.jfrog.io/afero/afero-java-sdk'
+		credentials {
+		    username = project.findProperty('aferoArtifactoryUserName') ?: ""
+		    password = project.findProperty('aferoArtifactoryPassword') ?: ""
+		}
+	   }
+       }
     }
-}
-
-javadoc {
-    options.addStringOption('Xdoclint:none', '-quiet')
-    options.links("https://square.github.io/okhttp/3.x/okhttp/")
-    options.links("https://square.github.io/retrofit/2.x/retrofit/")
-    options.links("http://fasterxml.github.io/jackson-annotations/javadoc/2.8/")
-    options.links("http://fasterxml.github.io/jackson-core/javadoc/2.8/")
-    options.links("http://fasterxml.github.io/jackson-databind/javadoc/2.8/")
-}
-
-task packageJavadoc(type: Jar) {
-    from javadoc
-    classifier = 'javadoc'
 }
