@@ -20,21 +23,25 @@ repositories {
     }
 }
 
-final String hubbyVersion = "1.0.957"
+final String hubbyVersion = "1.0.978"
 final String hubType = project.findProperty('aferoSofthubType') ?: '"CONSUMER"'
 
 android {
-    compileSdkVersion 30
+    namespace "io.afero.sdk.softhub"
+    compileSdkVersion 35
 
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 30
-
+        minSdkVersion 21
+        targetSdkVersion 35
         versionCode 1
-        versionName rootProject.version
+        versionName rootProject.version // Correct: Inherits version from root.
 
-        buildConfigField 'String', 'HUBBY_VERSION', '"'+hubbyVersion+'"'
+        buildConfigField 'String', 'HUBBY_VERSION', '"' + hubbyVersion + '"'
         buildConfigField 'String', 'HUB_TYPE', hubType
+    }
+
+    buildFeatures {
+        buildConfig true
     }
 
     buildTypes {
