@@ -31,29 +36,47 @@ configurations.all {
 
 
 android {
-    compileSdkVersion 32
+    namespace 'io.afero.aferolab'
+    compileSdkVersion 35
 
     defaultConfig {
         applicationId 'io.afero.aferolab'
         minSdkVersion 26
-        targetSdkVersion 32
+        targetSdkVersion 35
 
         versionCode 1
         versionName '1.0'
 
-        buildConfigField 'String', 'AFERO_CLIENT_ID', project.aferoClientId
-        buildConfigField 'String', 'AFERO_CLIENT_SECRET', project.findProperty('aferoClientSecret') ?: 'null'
-        buildConfigField 'String', 'AFERO_SERVICE_HOSTNAME', project.findProperty('aferoServiceHostname') ?: '"api.afero.io"'
-        buildConfigField 'String', 'AFERO_SOFTHUB_SERVICE', project.findProperty('aferoSofthubService') ?: 'null'
-        buildConfigField 'String', 'AFERO_SOFTHUB_AUTHENTICATOR_CERT', project.findProperty('aferoSofthubAuthenticatorCert') ?: 'null'
+        def getQuotedProp = { propName, defaultValue = null ->
+            def value = project.findProperty(propName)
+            if (value != null) {
+                return "\"${value}\""
+            }
+            if (defaultValue != null) {
+                return "\"${defaultValue}\""
+            }
+            return 'null'
+        }
 
-        buildConfigField 'String', 'AFERO_OAUTH_AUTH_URL', project.findProperty('aferoOauthAuthUrl') ?: 'null'
-        buildConfigField 'String', 'AFERO_OAUTH_TOKEN_URL', project.findProperty('aferoOauthTokenUrl') ?: 'null'
-        buildConfigField 'String', 'AFERO_OAUTH_REDIRECT_SCHEME', project.findProperty('aferoOauthRedirectScheme') != null ? '"' + project.findProperty('aferoOauthRedirectScheme') + '"' : 'null'
-        buildConfigField 'String', 'AFERO_OAUTH_REDIRECT_HOST',  project.findProperty('aferoOauthRedirectHost') != null ? '"' + project.findProperty('aferoOauthRedirectHost') + '"' : 'null'
+        buildConfigField 'String', 'AFERO_CLIENT_ID', getQuotedProp('aferoClientId')
+        buildConfigField 'String', 'AFERO_CLIENT_SECRET', getQuotedProp('aferoClientSecret')
+        buildConfigField 'String', 'AFERO_SERVICE_HOSTNAME', getQuotedProp('aferoServiceHostname', 'api.afero.io')
+        buildConfigField 'String', 'AFERO_SOFTHUB_SERVICE', getQuotedProp('aferoSofthubService')
+        buildConfigField 'String', 'AFERO_SOFTHUB_AUTHENTICATOR_CERT', getQuotedProp('aferoSofthubAuthenticatorCert')
 
-        manifestPlaceholders = [oAuthRedirectScheme: project.findProperty('aferoOauthRedirectScheme') != null ? project.findProperty('aferoOauthRedirectScheme') : "",
-                                oAuthRedirectHost: project.findProperty('aferoOauthRedirectHost') != null ? project.findProperty('aferoOauthRedirectHost') : ""]
+        buildConfigField 'String', 'AFERO_OAUTH_AUTH_URL', getQuotedProp('aferoOauthAuthUrl')
+        buildConfigField 'String', 'AFERO_OAUTH_TOKEN_URL', getQuotedProp('aferoOauthTokenUrl')
+        buildConfigField 'String', 'AFERO_OAUTH_REDIRECT_SCHEME', getQuotedProp('aferoOauthRedirectScheme')
+        buildConfigField 'String', 'AFERO_OAUTH_REDIRECT_HOST',  getQuotedProp('aferoOauthRedirectHost')
+
+        def aferoOauthRedirectScheme = project.findProperty('aferoOauthRedirectScheme') ?: ""
+        def aferoOauthRedirectHost = project.findProperty('aferoOauthRedirectHost') ?: ""
+	println 'XXXaferoOauthRedirectScheme "' + aferoOauthRedirectScheme + '"'
+	println 'XXXaferoOauthRedirectHost "' + aferoOauthRedirectHost + '"'
+        manifestPlaceholders = [
+                oAuthRedirectScheme: aferoOauthRedirectScheme,
+                oAuthRedirectHost  : aferoOauthRedirectHost
+        ]
 
     }
 
@@ -78,11 +101,16 @@ android {
         exclude 'META-INF/NOTICE.txt'
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
-
+    buildFeatures {
+        buildConfig true
+        viewBinding true
+    }
 }
+
+def camerax_version = "1.5.1"
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
