@@ -4,9 +4,6 @@ import android.content.Context;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.Button;
-import android.widget.ListView;
-import android.widget.TextView;
 
 import androidx.annotation.AttrRes;
 import androidx.annotation.NonNull;
@@ -14,13 +11,9 @@ import androidx.annotation.Nullable;
 import androidx.annotation.StringRes;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import butterknife.OnItemClick;
 import io.afero.aferolab.R;
+import io.afero.aferolab.databinding.ViewWifiSetupBinding;
 import io.afero.aferolab.widget.PasswordDialog;
-import io.afero.aferolab.widget.ProgressSpinnerView;
 import io.afero.aferolab.widget.ScreenView;
 import io.afero.sdk.client.afero.AferoClient;
 import io.afero.sdk.device.DeviceModel;
@@ -29,38 +22,7 @@ import rx.subjects.PublishSubject;
 
 public class WifiSetupView extends ScreenView {
 
-    @BindView(R.id.network_list_container)
-    View mListContainer;
-
-    @BindView(R.id.network_list_refresh)
-    SwipeRefreshLayout mListRefresh;
-
-    @BindView(R.id.network_list)
-    ListView mListView;
-
-    @BindView(R.id.network_list_empty_container)
-    View mEmptyListContainer;
-
-    @BindView(R.id.wifi_setup_title_label)
-    TextView mTitleText;
-
-    @BindView(R.id.wifi_setup_message_label)
-    TextView mMessageText;
-
-    @BindView(R.id.network_error)
-    View mErrorContainer;
-
-    @BindView(R.id.wifi_setup_progress)
-    ProgressSpinnerView mProgressView;
-
-    @BindView(R.id.wifi_setup_success)
-    View mWifiSetupSuccessView;
-
-    @BindView(R.id.wifi_send_creds_try_again_button)
-    Button mTryAgainSendWifiCredsButton;
-
-    @BindView(R.id.wifi_scan_try_again_button)
-    Button mTryAgainWifiScanButton;
+    private ViewWifiSetupBinding binding;
 
     private WifiSetupController mController;
     private PublishSubject<WifiSetupView> mViewSubject = PublishSubject.create();
@@ -85,7 +47,15 @@ public class WifiSetupView extends ScreenView {
     @Override
     public void onFinishInflate() {
         super.onFinishInflate();
-        ButterKnife.bind(this);
+        binding = ViewWifiSetupBinding.bind(this);
+
+        binding.refreshButton.setOnClickListener(this::onClickRefresh);
+        binding.emptyRefreshButton.setOnClickListener(this::onClickRefresh);
+        binding.networkList.setOnItemClickListener(this::onNetworkListItemClick);
+        binding.wifiErrorCancelButton.setOnClickListener(v -> onClickCancel());
+        binding.wifiScanTryAgainButton.setOnClickListener(v -> onClickWifiScanTryAgain());
+        binding.wifiSendCredsTryAgainButton.setOnClickListener(v -> onClickSendCredsTryAgain());
+        binding.wifiSetupDoneButton.setOnClickListener(v -> onClickDone());
     }
 
     public WifiSetupView start(DeviceModel deviceModel, AferoClient aferoClient) {
@@ -94,7 +64,7 @@ public class WifiSetupView extends ScreenView {
         mController = new WifiSetupController(this, deviceModel, aferoClient);
         mController.start();
 
-        mListRefresh.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+        binding.networkListRefresh.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
                 mController.onClickRefresh();
@@ -121,7 +91,7 @@ public class WifiSetupView extends ScreenView {
     }
 
     void setAdapter(WifiSSIDListAdapter adapter) {
-        mListView.setAdapter(adapter);
+        binding.networkList.setAdapter(adapter);
     }
 
     void askUserToTurnOnBluetooth(DeviceModel deviceModel) {
@@ -131,54 +101,54 @@ public class WifiSetupView extends ScreenView {
     }
 
     void showLookingProgress() {
-        mMessageText.setText(R.string.wifi_looking_for_device);
+        binding.wifiSetupMessageLabel.setText(R.string.wifi_looking_for_device);
 
-        mEmptyListContainer.setVisibility(GONE);
-        mListContainer.setVisibility(GONE);
-        mErrorContainer.setVisibility(GONE);
-        mProgressView.show();
+        binding.networkListEmptyContainer.setVisibility(GONE);
+        binding.networkListContainer.setVisibility(GONE);
+        binding.networkError.setVisibility(GONE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.VISIBLE);
     }
 
     void showConnectProgress() {
-        mMessageText.setText(R.string.wifi_connecting_to_device);
+        binding.wifiSetupMessageLabel.setText(R.string.wifi_connecting_to_device);
 
-        mEmptyListContainer.setVisibility(GONE);
-        mListContainer.setVisibility(GONE);
-        mErrorContainer.setVisibility(GONE);
-        mProgressView.show();
+        binding.networkListEmptyContainer.setVisibility(GONE);
+        binding.networkListContainer.setVisibility(GONE);
+        binding.networkError.setVisibility(GONE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.VISIBLE);
     }
 
     void hideProgress() {
-        mListRefresh.setRefreshing(false);
-        mProgressView.hide();
+        binding.networkListRefresh.setRefreshing(false);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.GONE);
     }
 
     void showEmptyView() {
-        mEmptyListContainer.setVisibility(VISIBLE);
-        mListContainer.setVisibility(GONE);
-        mProgressView.hide();
+        binding.networkListEmptyContainer.setVisibility(VISIBLE);
+        binding.networkListContainer.setVisibility(GONE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.GONE);
     }
 
     void showListView() {
-        mEmptyListContainer.setVisibility(GONE);
-        mListContainer.setVisibility(VISIBLE);
-        mErrorContainer.setVisibility(GONE);
-        mProgressView.hide();
+        binding.networkListEmptyContainer.setVisibility(GONE);
+        binding.networkListContainer.setVisibility(VISIBLE);
+        binding.networkError.setVisibility(GONE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.GONE);
     }
 
     void showWifiConnectProgress() {
-        mMessageText.setText(R.string.wifi_please_wait);
+        binding.wifiSetupMessageLabel.setText(R.string.wifi_please_wait);
 
-        mEmptyListContainer.setVisibility(GONE);
-        mListContainer.setVisibility(GONE);
-        mErrorContainer.setVisibility(GONE);
-        mProgressView.show();
+        binding.networkListEmptyContainer.setVisibility(GONE);
+        binding.networkListContainer.setVisibility(GONE);
+        binding.networkError.setVisibility(GONE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.VISIBLE);
     }
 
     void showWifiScanError() {
-        mMessageText.setText(R.string.wifi_cant_connect_to_device);
-        mTryAgainSendWifiCredsButton.setVisibility(GONE);
-        mTryAgainWifiScanButton.setVisibility(VISIBLE);
+        binding.wifiSetupMessageLabel.setText(R.string.wifi_cant_connect_to_device);
+        binding.wifiSendCredsTryAgainButton.setVisibility(GONE);
+        binding.wifiScanTryAgainButton.setVisibility(VISIBLE);
         showErrorContainer();
     }
 
@@ -187,26 +157,26 @@ public class WifiSetupView extends ScreenView {
     }
 
     void showSendWifiCredsError(@StringRes int resId) {
-        mMessageText.setText(resId);
-        mTryAgainSendWifiCredsButton.setVisibility(VISIBLE);
-        mTryAgainWifiScanButton.setVisibility(GONE);
+        binding.wifiSetupMessageLabel.setText(resId);
+        binding.wifiSendCredsTryAgainButton.setVisibility(VISIBLE);
+        binding.wifiScanTryAgainButton.setVisibility(GONE);
         showErrorContainer();
     }
 
     private void showErrorContainer() {
-        mEmptyListContainer.setVisibility(GONE);
-        mListContainer.setVisibility(GONE);
-        mErrorContainer.setVisibility(VISIBLE);
-        mProgressView.hide();
+        binding.networkListEmptyContainer.setVisibility(GONE);
+        binding.networkListContainer.setVisibility(GONE);
+        binding.networkError.setVisibility(VISIBLE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.GONE);
     }
 
     void showSuccess() {
-        mMessageText.setText(R.string.wifi_your_device_is_now_connected);
+        binding.wifiSetupMessageLabel.setText(R.string.wifi_your_device_is_now_connected);
 
-        mWifiSetupSuccessView.setVisibility(VISIBLE);
-        mListContainer.setVisibility(GONE);
-        mErrorContainer.setVisibility(GONE);
-        mProgressView.hide();
+        binding.wifiSetupSuccess.setVisibility(VISIBLE);
+        binding.networkListContainer.setVisibility(GONE);
+        binding.networkError.setVisibility(GONE);
+        binding.wifiSetupProgress.getRoot().setVisibility(View.GONE);
     }
 
     void onCompleted() {
@@ -214,32 +184,26 @@ public class WifiSetupView extends ScreenView {
         mViewSubject.onCompleted();
     }
 
-    @OnClick({ R.id.refresh_button, R.id.empty_refresh_button })
     void onClickRefresh(View view) {
         mController.onClickRefresh();
     }
 
-    @OnItemClick(R.id.network_list)
     void onNetworkListItemClick(AdapterView<?> parent, View view, int position, long id) {
         mController.onNetworkListItemClick(position);
     }
 
-    @OnClick(R.id.wifi_error_cancel_button)
     void onClickCancel() {
         mController.onClickCancel();
     }
 
-    @OnClick(R.id.wifi_scan_try_again_button)
     void onClickWifiScanTryAgain() {
         mController.onClickWifiScanTryAgain();
     }
 
-    @OnClick(R.id.wifi_send_creds_try_again_button)
     void onClickSendCredsTryAgain() {
         mController.onClickWifiConnectTryAgain();
     }
 
-    @OnClick(R.id.wifi_setup_done_button)
     void onClickDone() {
         onCompleted();
     }
